Verify wheel
