+++ bindings/py/cpp_src/CMakeLists.txt
@@ -102,8 +102,8 @@ source_group("test"       FILES ${src_py_test_files})
 # setup the distr directory
 set(distr ${CMAKE_INSTALL_PREFIX}/distr)
 file(MAKE_DIRECTORY ${distr})
-file(COPY ../packaging/. DESTINATION ${distr} PATTERN *)
-file(COPY ${REPOSITORY_DIR}/py/. DESTINATION ${distr}/src PATTERN *)
+file(COPY "../packaging/*" DESTINATION ${distr} REGEX  ".*__pycache__" EXCLUDE)
+file(COPY ${REPOSITORY_DIR}/py/. DESTINATION ${distr}/src REGEX  ".*__pycache__" EXCLUDE)
 
 # determine which version of binding to build
 if ("${BINDING_BUILD}" STREQUAL "Python2")
