+++ bindings/py/packaging/src/tests/regions/network_test.py
